requirements.txt
absl-py==1.4.0
astroid==2.14.1
astunparse==1.6.3
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
-e git+https://github.com/alicedoherty/bayesian-individual-fairness.git@9734aea2609f0cbef24e8e4896a102f2d839e0f3#egg=deepbayes&subdirectory=deepbayes
dill==0.3.6
dm-tree==0.1.8
flatbuffers==23.1.21
fonttools==4.39.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
isort==5.12.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.5.0
platformdirs==2.6.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pylint==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
statsmodels==0.13.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.0
